+++ src/util/shared/builtin.ts
@@ -1,0 +1,4 @@
+/**
+ * Represents values that can be serialize / deserialized via `JSON.stringify` / `JSON.parse`
+ */
+export type IJSONType = boolean | number | string | IJSONType[] | {[key: string]: IJSONType};