+++ types/stores/shared/index.d.ts
@@ -1,3 +1,4 @@
 export * from "./collection";
 export * from "./immutable";
 export * from "./merged";
+export * from "./schema";