+++ types/util/shared/index.d.ts
@@ -1,4 +1,5 @@
 export * from "./actions";
+export * from "./builtin";
 export * from "./browser";
 export * from "./context";
 export * from "./functional";